Added support for +,-,*,/ as binary operators to reference syntax
Allow the basic arithmetic operators to be invoked using infix syntax.

There's a bit of oddness here WRT how precedence for these works with
functions.  As it stands, "foo bar + baz bot" is treated as "foo(bar + baz,
bot)" which doesn't make a lot of sense.  May be better to treat functions as
higher precedence.