Fix bug in the way lexical contexts are referenced.
Storing the lexical context of an object as a dictionary with a parent
reference causes a cycle when the same context is used multiple times on the
stack.

Break out the stack functionality (the "parent" pointer) from the variable
dictionary functionality so we can simply create a new, lightweight lexical
context stack frame when entering a method reference.