Implement new-style methods and additional implicit parameters.
New style methods don't curry the receiver as the existing methods do.
Instead, they extract it from the eval context.  This means that they can be
defined once on a prototype object that is the parent of all instances instead
of redefined for every instance.  These are defined using the new `@method`
annotation.

We also define two special keywords that can be included in `@function` or
`@method` annotation parameters: "wrap" and "ctx".  If "wrap" is specified, the
function gets an implicit "wrap" variable that is the ObjOf[] wrapper around
the function (this is mainly useful for defining cascade-style methods that
return the receiver).  If "ctx" is specified, the function gets an implicit
"ctx" variable set to the eval context.